+++ src/DataObjects/CheckoutSessionData.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Subster\PhpSdk\DataObjects;
 
 use Subster\PhpSdk\Concerns\Data;
+use Subster\PhpSdk\Enums\CheckoutPaymentAttemptState;
 use Subster\PhpSdk\Enums\CheckoutPaymentState;
 
 class CheckoutSessionData extends Data
@@ -14,6 +15,9 @@ class CheckoutSessionData extends Data
         public readonly string $url,
         public readonly CheckoutPaymentState $payment_state = CheckoutPaymentState::RequiresPayment,
         public readonly ?string $checkout_url = null,
+        public readonly ?CheckoutPaymentAttemptState $payment_attempt_state = null,
+        public readonly ?float $amount = null,
+        public readonly ?string $currency = null,
     ) {}
 
     public static function fromSaloon(array $response): self
@@ -29,6 +33,11 @@ class CheckoutSessionData extends Data
                 ? CheckoutPaymentState::from(strval($response['payment_state']))
                 : CheckoutPaymentState::RequiresPayment,
             checkout_url: $checkoutUrl,
+            payment_attempt_state: isset($response['payment_attempt_state'])
+                ? CheckoutPaymentAttemptState::from(strval($response['payment_attempt_state']))
+                : null,
+            amount: isset($response['amount']) ? floatval($response['amount']) : null,
+            currency: isset($response['currency']) ? strval($response['currency']) : null,
         );
     }
 }